Move benchmarks to an independent project
* Benchmarks now depend on ACCP and there is no need to build ACCP
* Benchmarks can be run against published ACCP artifacts on Maven
