+++ benchmarks/.gitignore
@@ -1,0 +1,8 @@
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
+lib/build
+.idea
+.attach_pid*
