+++ benchmarks/lib/src/jmh/resources/com/amazon/corretto/crypto/provider/testdata/hmac.txt
@@ -1,0 +1,1 @@
+../../../../../../../../../../../extra-jar-files/com/amazon/corretto/crypto/provider/testdata/hmac.txt
