+++ benchmarks/settings.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    // Apply the foojay-resolver plugin to allow automatic download of JDKs
+    id("org.gradle.toolchains.foojay-resolver-convention") version "0.4.0"
+}
+
+rootProject.name = "benchmarks"
+include("lib")
